Create print string between brackets
Write a program to input a phrase that is guaranteed to contain an opening
parenthesis and a closing parenthesis, in that order, and is possibly repeated. For
example, such a phrase might be
for (int index=0; index != 5; index++) screen.write(index);
 Scan each character in the phrase and output only those characters that are
contained between the opening and closing parentheses.
 Rescan the phrase and output only those characters that are outside of the
opening and closing parentheses.